+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # ActFramework Change Log
 
-**1.8.23**
+**1.8.23** 16/May/2019
 * CLI Table view - print line interlaced with revert color/bg color #1143
 * Add `@Label` annotation for bean field marking #1142
 * Email validation handler - not able to valid email at `thinking.studio` #1141